Eliminate adding indicator functions, implement custom functions.
Composition of prox operators for added indicator functions is not
correct in general. It works for the one case I was using it for
(l1-norm + zeros indicator), and I had imagined that it held in more
general circumstances. This is not true, so remove it.

Sort of as a replacement, add a CustomFunction function that creates a
FunctionWithGradProx class to hold a given function, gradient, and prox
operator defined as separate python functions.